requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
dj-database-url==2.1.0
dj-static==0.0.6
Django==4.2.8
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
h11==0.14.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
ndg-httpsclient==0.5.1
openapi-codec==1.3.2
packaging==24.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pusher==3.3.2
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
#urllib3==2.1.0
urllib3==1.26.6
uvicorn==0.29.0
whitenoise==6.6.0
